utils.id_generator methods moved to legacy.utils
