+++ virtool_cli/ref/legacy.py
@@ -7,9 +7,7 @@ from typing import Any
 
 import click
 
-from virtool_cli.utils.id_generator import (
-    generate_random_alphanumeric,
-)
+from virtool_cli.legacy.utils import generate_random_alphanumeric
 from virtool_cli.utils.deprecated.reference import generate_otu_dirname, is_v1
